+++ CHANGELOG.md
@@ -410,3 +410,5 @@ This update overhauls Predator's dashcam functionality and adds various features
     - Thread names are now automatically determined and assignment when debug messages are displayed.
 - Fixed speed display in real-time mode.
 - Improved GPS handling.
+    - 2D position information can now be displayed even when a 3D fix has not been aquired yet.
+- Updated the time handling back-end to allow for custom time offsets.